Fix all tests except for differentiability ones

+++ cheetah/nocelot.py
@@ -28,25 +28,31 @@ def ocelot2cheetah(element, warnings: bool = True) -> "cheetah.Element":
         )
 
     if isinstance(element, ocelot.Drift):
-        return cheetah.Drift(length=torch.tensor(element.l), name=element.id)
+        return cheetah.Drift(
+            length=torch.tensor(element.l, dtype=torch.float32), name=element.id
+        )
     elif isinstance(element, ocelot.Quadrupole):
         return cheetah.Quadrupole(
-            length=torch.tensor(element.l), k1=torch.tensor(element.k1), name=element.id
+            length=torch.tensor(element.l, dtype=torch.float32),
+            k1=torch.tensor(element.k1, dtype=torch.float32),
+            name=element.id,
         )
     elif isinstance(element, ocelot.Hcor):
         return cheetah.HorizontalCorrector(
-            length=torch.tensor(element.l),
-            angle=torch.tensor(element.angle),
+            length=torch.tensor(element.l, dtype=torch.float32),
+            angle=torch.tensor(element.angle, dtype=torch.float32),
             name=element.id,
         )
     elif isinstance(element, ocelot.Vcor):
         return cheetah.VerticalCorrector(
-            length=torch.tensor(element.l),
-            angle=torch.tensor(element.angle),
+            length=torch.tensor(element.l, dtype=torch.float32),
+            angle=torch.tensor(element.angle, dtype=torch.float32),
             name=element.id,
         )
     elif isinstance(element, ocelot.Cavity):
-        return cheetah.Cavity(torch.tensor(element.l), name=element.id)
+        return cheetah.Cavity(
+            torch.tensor(element.l, dtype=torch.float32), name=element.id
+        )
     elif isinstance(element, ocelot.Monitor) and ("BSC" in element.id):
         # NOTE This pattern is very specific to ARES and will need a more complex
         # solution for other accelerators
@@ -63,14 +69,18 @@ def ocelot2cheetah(element, warnings: bool = True) -> "cheetah.Element":
     elif isinstance(element, ocelot.Monitor) and "BPM" in element.id:
         return cheetah.BPM(name=element.id)
     elif isinstance(element, ocelot.Undulator):
-        return cheetah.Undulator(torch.tensor(element.l), name=element.id)
+        return cheetah.Undulator(
+            torch.tensor(element.l, dtype=torch.float32), name=element.id
+        )
     else:
         if warnings:
             print(
                 f"WARNING: Unknown element {element.id} of type {type(element)},"
                 " replacing with drift section."
             )
-        return cheetah.Drift(length=torch.tensor(element.l), name=element.id)
+        return cheetah.Drift(
+            length=torch.tensor(element.l, dtype=torch.float32), name=element.id
+        )
 
 
 def subcell_of_ocelot(cell: list, start: str, end: str) -> list:
